destroy action refactored to work with Vue and working

+++ app.py
@@ -56,12 +56,14 @@ def update(todo_id):
     return jsonify({"error": "Title not provided"}), 400
 
 
-@app.route("/delete/<int:todo_id>")
+@app.route("/delete/<int:todo_id>", methods=["DELETE"])
 def delete(todo_id):
     todo = Todo.query.filter_by(id=todo_id).first()
-    db.session.delete(todo)
-    db.session.commit()
-    return redirect(url_for("home"))
+    if todo:
+        db.session.delete(todo)
+        db.session.commit()
+        return jsonify({"message": "Todo Destroyed"}), 200
+    return jsonify({"error": "Todo not found"}), 404
 
 
 def create_tables():
